arb program: Add trig instruction template
This template is used to verify COS, SCS, and SIN instructions.  This
is something of a cross between binary_result.template and
mustpass.template.  The binary pass / fail result is written to the
red channel of the color output.  The raw result of the instruction is
remapped to [0,1] and stored in the green and blue channels.

Since the result of these trig instructions is always on the range
[-1,1] there is no need to calculate a tolerance value for each set of
inputs.  Instead, a hard-coded value of 1e-4 is used.

The program for the cos.shader_test for ARB_fragment_program will look
like:

!!ARBfp1.0
	TEMP	R0, R1;

	MOV	R0, program.env[1].xxxx;
	COS	R0, program.env[0].x;

	# If the value in R0 is not within the tolerance of the expected
	# result (program.env[3]) set the red component to 1.0.
	SUB	R1, R0, program.env[3].xyxy;
	ABS	R1, R1;
	SLT	R1, R1, {1e-4}.x;		# Abs diff less than tolerance?
	MUL	R1.xy, R1.xzxz, R1.ywyw;	# R1.xy = ((R1.x && R1.y),
						#          (R1.z && R1.w))
	MUL	R1.x, R1.x, R1.y;		# R1.x = (R1.x && R1.y
						#         && R1.z && R1.w)
	SUB	R1.y, {1.0}.x, R1.x;		# Logical not

	MUL	result.color.x, R1.y, {1.0}.x;

	# Store the X and Y components of R0 in the green and blue components
	# of the output.  Compress the range of the values from [-1,1] to
	# [0,1].  Also, force W to be 1.
	MAD	result.color.yzw, R0.xxyy, {0.0, 0.5, 0.5, 0.0}, {0.0, 0.5, 0.5, 1.0};
	END

Signed-off-by: Ian Romanick <ian.d.romanick@intel.com>