Run random UBO tests forever looking for failures
Doesn't actually run forever.  Passing tests are put in a directory
called "pass", and failing tests are put in a directory called "fail".
The test will terminate when 1,000 failing tests are generated.

This is an example of how random_ubo.py can be used.  This is how I
found SIX bugs in Mesa's UBO code.  I don't think we want this in
piglit, but it should live somewhere.

Signed-off-by: Ian Romanick <ian.d.romanick@intel.com>
Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>