gen_tess_input_tests.py: use numpy for random data
Python 2.x and python 3.x produce different results. This is probably a
difference between the implementations. This result is largely the same
in order of magnitude differences, but the numbers themselves are
sometimes significantly different. This might be due to the difference
between python 2 and python 3's int implementations (int in python 2 can
roll over, but it will grow to consume all memory in python 3), or it
might be due to differences in random.

When ported to numpy both python 2.x and python 3.x produce the same
result, which is very close to the value that python 3.x produces.
This produces deterministic results between python 2.x and python 3.x.

This has not been tested, i965 lacks tessalation support.