generated_tests/gen_tess_input_tests.py: Reduce code duplication
This uses a couple of mixin classes to reduce code duplication. It
splits the identical 'generate' functions out into a mixin, as well as
the 'components' and 'test_data' methods into a second mixin. It then
makes filename a lazy_property, to make even less code.

The result is the removal of about 50 LOC.