deqp: Add a group run mode for dEQP.
Currently dEQP integration works by running one test at a time. This
provides good process isolation and stability, but at the cost of a good
deal of overhead to start each test. dEQP was not designed to be run one
test at a time, for all of it's tests. This overhead can obviously be
overcome by invoking dEQP directly which overcomes the performance
problem, but that has it's own problems.

This patch attempts to strike a balance between performance and
reliability by adding a new mode that runs groups of tests in a single
process. This works by finding each leaf node in the dEQP group
structure and running that as a single test process, and uses subtests
to report the result.

For piglit summary tools this is indistinguishable (although the
internal JSON representation is different), except for some metadata
(run time, stdout and stderr, etc), but the regressions/fixes/ etc will
be the same.

In my tests this reduces run time of the test suite between ~30% and
~50%.

deqp-vk can run in group mode, but doesn't have the skip conditionals.
That will be addressed in a follow up patch. There is also the issue
that the JUnit backend doesn't handle subtests. This is an ongoing
problem to solve.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>