deqp: in group mode, rerun tests that fail in test at a time mode
This adds a new feature to the deqp integration group at a time mode,
which allows tests to be rerun when they meet a set of bad status in
test at a time mode. This helps to eliminate tests where the crash
caused tests to not be run, and helps clear up issues where an error
would change test statuses.

In my runs this is still significantly faster than test at a time mode,
unless a huge number of tests are crashing. In that case it is really up
to the developer to know whether they need test at a time mode, or
whether group at a time mode is sufficient.

A new option, --deqp-no-group-rerun, is added which disables this
feature.

This also addresses the vulkan suite issue, by rerunning any tests where
the problematic statuses are detected in test at a time mode.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>