framework: remove unused keyword argument from Testprofile.group_manager
This extra keyword argument (prefix) is completely optional, and isn't
used and never was. Since it's a keyword it doesn't affect the function
signature for callers that didn't use it, and since it is unused no
function should specify it. It can also be added back if it ever becomes
useful.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>