framework: remove unused backends method
Once upon a time piglit stored all of its results throughout the entire
test run. This is a bad idea of course, and piglit doesn't do it
anymore. However, there was some legacy code in the backend (and run.py)
for setting metadata on this TestResult instance. This patch removes all
of that code.k