shader_runner: Allow whitespace before [require] declarations
Currently any leading whitespace in a line of the require section
results in that line being skipped. This is less than optimal because of
generated tests that may want to make their templates more readable at
the expense of having some odd indent in the produced shader.

This is required to get the new, faster templates later in this series
working.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>