framework: Make some arguments per profile.
This breaks the command line syntax for piglit in a non-backwards
compatible way. I'd really prefer not to muck with it for my own sanity
as much as anything else, but I don't see how to fix it without
breaking it.

The problem is that piglit conflates runner arguments with profile
arguments, and this series aims to fix profiles in piglit, because
multiple profiles is inherently broken. The goal of this patch is to
allow each profile to specify it's own arguments in conjunction with
using a set of standard arguments.

Further patches in this series will continue to de-tangle profiles from
each other and make this more valuable.

TODO: this still needs to be plugged in to the lower levels
TODO: this probably needs some testing