framework: Add an initial implementaiton of a DEQPGroupTest
This patch adds a new DEQP base class that uses the ReducedProcessMixin
to provide a "group at a time" mode, in which a group leaf of the DEQP
test hierarchy is run in a single process. It doesn't remove the
existing "test at a time" mode, where a single test is runin a single
process. It also provides two classes built on top of that class meant
to be used by concrete implementations.  One that uses asterisks and one
that uses trie lists. Each have their advantages. The asterisks are
simpler and much shorter, the trie lists can represent groups that have
both tests and groups, but can grow larger than dEQP can parse.

The implementation details are such that some shared bits are split out
into a DEQPBaseTest, and a DEQPSingleTest and DEQPGroupTest implement
the two modes. This renames the old DEQPBaseTest to DEQPSingleTest,
which requires touching all of the deqp suites to change their parent
class.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>