framework: Fix FastSkipMixin version comparisons
Currently the FastSkipMixin assumes that versions will always be a
minimum required version, and nothing else. This is mostly true of
piglit tests, but isn't always true. Sometimes tests specify that they
need a version greater-or-equal, and a few have less-than or
less-than-equal.

This patch fixes that, and also generalizes the fast-skip mechanism a
bit. It does this by allowing the caller to set an operation as well as
a value.

This also updates glslparsertest and shadertest to use the newly updated
interfaces.

v2: - fix the comparison for >= and <=

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>