[RFC] framework: Use an enum for concurrency.
This uses an enum (which are new in python 3.4, but available via the
enum34 backport in pip). I'm putting this up as an RFC, since it does
introduce another dependency for python 2.7 and 3.3, but I think it
makes the code cleaner, since this doesn't require passing strings
around for controlling concurrency.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>