framework: Add a Register metaclass for Test classes
This creates a registry of tests that will be used when deserializing XML
back into object representations of tests.

A large number of the changes in this patch are in unittests, which now
fail because they assume that they create duplicate named Test derived
classes, and assumption which is no longer true without mocking.