framework: Add configuration option for format of names in expected
In current master only JUnit escaped names are supported, and this is in
active use by Intel. However, this is less than optimal and ideally
would be deprecated and unsupported in favor of using native piglit
names. This patch allows the old behavior to continue working, but
defaults to the new behavior.

TODO: tests