framework: Split Test.interpret_result into a Mixin
This method is used by some, but not all implementations. It doesn't
makes sense to have a shared bas class implement code that isn't used by
half of it's children, splitting it into a Mixin class allows the
children that need this code to include it if they want it, but ignore
it otherwise.