opengl.py: Remove exported gl extensions
This conflicts with the idea of serializing the tests at build time,
since we can't assume that the build machine has the same features set
as the machine that will run tests, we can't run such a filter.

Since startup time is much faster and fast skip calculations are done at
build time I hope that no one will miss this.