tests/quick: fix filtering of vs_in shader tests
This filter is incorrect as is, in that it can mask tests that it can
remove tests it shouldn't, for example
spec@arb_enhanced_layouts@arb_enhanced_layouts-transform-feedback-layout-qualifiers_vs_interface.
Correcting this is problematic because it changes the way the filter
works and filter will now remove different tests than it did before.

However, it is necessary for a couple of reasons. 1) it's removing tests
it shouldn't, 2) this series is going to split reverse the
shader.py/all.py relationship, and this bug is going to surface there.

The good new is that while this changes the specific tests run, at least
on Intel hardware no failures are removed and no failures are added.